// src/types/button.d.ts
export type ButtonType =
  | "primary"
  | "primary-outline"
  | "success-outline"
  | "warning-outline"
  | "danger-outline"
  | "success"
  | "warning"
  | "danger"
  | "outline"
  | "default";

export type ButtonSize = "xsmall" | "small" | "medium" | "large";

export type ButtonShape = "default" | "rounded" | "flat";